feat: Add compile, deploy and signining of a zkApp
Adds the functionality to compile, deploy and sign a zkApp. In order to
compile and deploy, we need to get an instance of the SmartContract
class that the user wants to deploy. To get a class instance in the
zkapp-cli, we look for the specified contract to deploy and add a
`export default` statement to the class declaration. This way, we can
easily import the class definition into our compile and deploy workflow.
We use a temporary build file to do this as to not change any expected
behavior of building the typescript files and cause confusion.