feat: Refactor regex logic to object lookups
This commit refactors the regex string checking for export statements
and instead puts the responsibility on the user to add the export
statements. If we do not find the expected class that the user specifies
in their config.json, we log an error message and tell them to change
it.