feat(ghp-postbuild.js): add script to modify built CSS files and prepend repo-name to asset URLs for compatibility with GitHub Pages deployment
The `ghp-postbuild.js` script is added to modify the built CSS files and prepend the repository name to the asset URLs. This is necessary to ensure compatibility with GitHub Pages deployment. The script reads the CSS files in the `.next/static/css` directory and replaces all occurrences of `url(/)` with `url(/<repoURL>/)`, where `<repoURL>` is the repository name specified in the script. This ensures that the asset URLs are correctly resolved when the app is deployed on GitHub Pages.