feat(deploy.js): add function findZkProgramFile to find the file name of the user-specified ZkProgram
The function `findZkProgramFile` is added to the `deploy.js` file. This function takes in two parameters: `buildPath` and `zkProgramName`. It uses a glob pattern to find all files matching the `buildPath` pattern. It then searches for a specific pattern in each file to find the user-specified zkprogram. If a match is found, it returns the file name of the zkprogram.

This function is useful for finding the file name of the zkprogram that needs to be compiled in the deployment process.