feat(preconditions.ts): add ElectionPreconditions and ParticipantPreconditions classes
The ElectionPreconditions class is added to represent the preconditions for an election. It has properties for the start and end of the election as UInt32 values, and a boolean flag to indicate if the preconditions should be enforced. The default preconditions are set to startElection = 0, endElection = MAXINT, and enforce = false.

The ParticipantPreconditions class is added to represent the preconditions for a participant in the election. It has properties for the minimum and maximum Mina values as UInt64 values. The default preconditions are set to minMina = 0 and maxMina = MAXINT.