feat(deploy.js): add getZkProgram function to retrieve zk program from build path and project root
The getZkProgram function is added to the deploy.js file. This function is responsible for retrieving the zk program from the specified build path and project root. It takes in the build path, project root, and zk program name as arguments. This function will be used to find the file name of the smart contract to be deployed.